// src/app/[lang]/[appId]/(web)/news-detail/[id]/interface.ts
export interface INewItem {
  id: number;
  title: string;
  status: string;
  createdAt: string;
  updatedAt: string;
  thumbnail: {
    id: number;
    url: string;
    type: string;
  };
  subject: {
    id: number;
    priority: number | null;
    subjectDetails: {
      id: number;
      lang: string;
      name: string;
    }[];
  }[];
  newsDetails: {
    id: number;
    lang: string;
    content: string;
    description: string;
    author: string;
  }[];
}

export interface IListNews {
  items: INewItem[];
  meta: {
    totalItems: number;
    itemCount: number;
    itemsPerPage: number;
    totalPages: number;
    currentPage: number;
  };
}